requirements.txt
asgiref==3.8.1
cffi==1.17.1
contourpy==1.3.1
cryptography==40.0.2
cycler==0.12.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.10
fonttools==4.57.0
inflection==0.5.1
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mysqlclient==2.2.7
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.5.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.1
uritemplate==4.1.1
xgboost==3.0.0
